authenticate and authorize user using jwt and delete posts

+++ controllers/api/v1/posts_api.js
@@ -22,17 +22,18 @@ module.exports.index = async function(req,res){
 module.exports.destroy = async function(req,res){
     try {
         let post = await Post.findById(req.params.id);
-        // if(post.user == req.user.id){
+        if(post.user == req.user.id){
             post.remove();
             await Comment.deleteMany({post: req.params.id});
 
            return res.json(200,{
                message:"post and associated comments deleted successfully"
            });
-        // }else{
-        //     req.flash('error','you cannot delete this post');
-        //     return res.redirect('back');
-        // }    
+        }else{
+           return res.status(401).json({
+               message: 'you cannot delete this post'
+           });
+        }    
     } catch (error) {
         return res.json(500,{
             message: "internal server error"
